Issue #2: Allow copy-resources to take in multiple directory arguments
Previously the options.copy.src could only take in a single
configuration. This means that you must have ALL resources under a
single directory, which can get messy. In my change I allow you to
configure the copy task with an array of source locations.

This is achieved by creating the previous pairing array for each
item in the copy.src array and then flattening the resulting nested
array.

Note:
The configured excludes will be applied to ALL items in the configured
source array.